Refactoring for Part Two
Part One tests passing

+++ 2023/csharp/Advent23.Days/Day11/Galaxy.cs
@@ -11,6 +11,6 @@
     {
         public required int Source { get; init; }
         public required int Target { get; init; }
-        public required int Steps { get; init; }
+        public required long Steps { get; init; }
     }
 }
